basic user mouse input

+++ src/main/java/org/openteufel/game/EntityManager.java
@@ -63,7 +63,7 @@ public class EntityManager
                 final long diffX = Math.abs(ent.getPosX() - x);
                 final long diffY = Math.abs(ent.getPosY() - y);
                 if (diffX > retdist || diffY > retdist)
-                    continue;
+                    continue; // avoid sqrt if possible
                 final double dist = Math.sqrt(diffX * diffX + diffY * diffY);
                 if (dist < retdist)
                 {
@@ -82,10 +82,10 @@ public class EntityManager
         {
             if (ent != null)
             {
-                final int diffX = Math.abs(ent.getPosX() - x);
-                final int diffY = Math.abs(ent.getPosY() - y);
+                final long diffX = Math.abs(ent.getPosX() - x);
+                final long diffY = Math.abs(ent.getPosY() - y);
                 if (diffX > radius || diffY > radius)
-                    continue;
+                    continue; // avoid sqrt if possible
                 final double dist = Math.sqrt(diffX * diffX + diffY * diffY);
                 if (dist <= radius)
                     ret.add(ent);
